+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.2](https://github.com/Howard86/next-api-handler/compare/v0.0.1...v0.0.2) (2021-10-17)
+
+
+### Features
+
+* **middleware:** add use method to inject middleware to req ([0b4a11e](https://github.com/Howard86/next-api-handler/commit/0b4a11eba7748d2a0730cb8c1b5830b5165113a9))
+
+
+### Bug Fixes
+
+* **config:** fix ava unit test config with serial ([7c8c59f](https://github.com/Howard86/next-api-handler/commit/7c8c59ffc6f0bcb84bd48d9665bbc73c16dddadb))
+
 ### 0.0.1 (2021-10-08)